chore(experiment): compose SAL-0032A
