+++ apps/observer/src/internal.ts
@@ -31,9 +31,9 @@ export * from "./runtime/eventBus.js";
 export * from "./runtime/logging.js";
 export * from "./runtime/main.js";
 export * from "./runtime/observerBootClaim.js";
-export * from "./runtime/observerDuplicateCleanup.js";
 export * from "./runtime/observerHandoff.js";
 export * from "./runtime/observerProcessEvidence.js";
+export * from "./runtime/observerReap.js";
 export * from "./runtime/reconcileScheduler.js";
 export * from "./runtime/server.js";
 export * from "./sqlite.js";
