+++ docs/debugging.md
@@ -323,11 +323,11 @@ of merging different nonempty provider databases. Override provider locations wi
 `--target-codex-home`, `--target-opencode-db`, and
 `--target-claude-projects` when the target uses isolated homes.
 
-After startup reconcile, the Observer performs one report-only duplicate
+After startup reconcile, the Observer performs one read-only duplicate
 inspection. `stn doctor` reports this as `observer-singleton`: an eligible
 candidate or evidence refusal is a warning, while a clear result is healthy.
-The structured Observer log records `would-terminate`, candidate PIDs, refusal
-codes, keeper preservation, and claim release. Report-only mode sends no signal.
+The structured Observer log records candidate PIDs and refusal evidence. The
+inspection has no signal or boot-claim authority.
 Use `stn observer reap` to compare the current process, holder, pidfile,
 socket-identity, start-token, launch-token, build, executable, argv, and complete
 Unix-socket-FD evidence.
@@ -335,9 +335,8 @@ Unix-socket-FD evidence.
 `stn observer reap --force` remains the explicit operator path for a confirmed
 duplicate. It holds the boot claim, refreshes health and strict process/FD evidence,
 sends SIGTERM, and may send SIGKILL after the manual grace period only after another
-fresh revalidation. Automatic cleanup has no SIGKILL path: terminate mode, when
-promoted separately after field evidence, sends one SIGTERM and reports a
-survivor for manual inspection. Do not use forced reap as a generic response to
+fresh revalidation. Station has no automatic duplicate-process signal path. Do not
+use forced reap as a generic response to
 an inaccessible socket or a live wedged owner.
 
 `OBSERVER_BUILD_MISMATCH` means a client outlived the exact Observer selector
