+++ tests/diagnostics/boundary-inventory.test.ts
@@ -78,8 +78,8 @@ const setTimeoutAllowlist = new Map([
     "Station host requests use per-request socket timeouts at the host protocol boundary, outside shared observer runtime helpers.",
   ],
   [
-    "apps/observer/src/runtime/observerDuplicateCleanup.ts",
-    "Duplicate quarantine and SIGTERM or explicit SIGKILL grace delays are cancellable OS process timing, not command timeout or retry plumbing.",
+    "apps/observer/src/runtime/observerReap.ts",
+    "Explicit SIGTERM and SIGKILL grace delays are operator-authorized OS process timing, not command timeout or retry plumbing.",
   ],
   [
     "apps/cli/src/observerProcess/startup.ts",
